requirements.txt
fastapi==0.68.0
pymongo==3.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
uvicorn==0.15.0
redis==4.0.2
motor==2.5.1
pydantic[email]==1.8.2  # Updated line (version might vary)